refactor: decompose common/utils.py into logical modules [BLDX-998]
common/utils.py was a 593-line dumping ground with 14 unrelated functions.
Decomposed into focused modules:

New modules:
- common/sql_filters.py — SQL filter pipeline (8 functions: prepare_query,
  prepare_filters, normalize_filters, parse_filter_input, get_database_names,
  extract_database_names_from_regex_common, transform_posix_regex, read_sql_files)
- credentials/utils.py — parse_credentials_extra
- common/concurrency.py — get_actual_cpu_count, get_safe_num_threads

Deleted (dead code):
- has_custom_control_config — 0 callers
- get_file_names — 0 callers, v2 legacy replaced by storage.batch

Remaining in common/utils.py (44 lines):
- download_file_from_upload_response — test-only file endpoint helper

All 2099 tests passing. Import paths updated across all callers.
