+++ application_sdk/clients/sql.py
@@ -29,8 +29,8 @@ from application_sdk.common.aws_utils import (
 )
 from application_sdk.common.error_codes import ClientError, CommonError
 from application_sdk.common.exc_utils import rewrap
-from application_sdk.common.utils import parse_credentials_extra
 from application_sdk.constants import AWS_SESSION_NAME, USE_SERVER_SIDE_CURSOR
+from application_sdk.credentials.utils import parse_credentials_extra
 from application_sdk.observability.logger_adaptor import get_logger
 
 logger = get_logger(__name__)
